requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.28.83
botocore==1.31.83
certifi==2023.7.22
charset-normalizer==3.2.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-ckeditor==6.7.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-parler==2.3
django-redis==5.3.0
django-rest-swagger==2.2.0
django-storages==1.14
django-taggit==4.0.0
django-tinymce==3.6.1
djangorestframework==3.14.0
drf-yasg==1.21.7
gunicorn==21.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
liqpay-python==1.0
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
requests==2.31.0
s3transfer==0.7.0
sentry-sdk==1.34.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
watchtower==3.0.1
django-prometheus==2.4.1
